weather/requirements.txt
functions-framework
SQLAlchemy
PyMySQL
pandas
pysqlite3-binary
requests
bs4
lat_lon_parser